+++ lib/wp_json_api.dart
@@ -19,15 +19,28 @@ import 'package:flutter/widgets.dart';
 import 'package:wp_json_api/helpers/typedefs.dart';
 import 'package:wp_json_api/networking/network_manager.dart';
 
+/// The base class to initialize and use WPJSONAPI
 class WPJsonAPI {
+
+  /// Private constructor for WPJSONAPI
   WPJsonAPI._privateConstructor();
 
+  /// Instance of WPJSONAPI
   static final WPJsonAPI instance = WPJsonAPI._privateConstructor();
 
+  /// The base url for the WordPress Site e.g. https://mysitewp.com
   String _baseUrl;
+
+  /// Debug boolean for outputting to the log
   bool _shouldDebug;
+
+  /// Default API root for your WordPress site
   String _apiPath = "/wp-json";
 
+  /// Initialize and configure class interface.
+  /// You can optional set [shouldDebug] == false to stop debugging
+  /// [wpJsonPath] is the root path for accessing you sites WordPress APIs
+  /// by default this should be "/wp-json".
   initWith(
       {@required String baseUrl,
       String wpJsonPath = '/wp-json',
@@ -37,27 +50,34 @@ class WPJsonAPI {
     _setShouldDebug(value: shouldDebug);
   }
 
+  /// Sets the base API in the class
   _setBaseApi({@required baseUrl}) {
     this._baseUrl = baseUrl;
   }
 
+  /// Sets the API path in the class
   _setApiPath({@required path}) {
     this._apiPath = path;
   }
 
+  /// Sets the debug value in the class
   _setShouldDebug({bool value}) {
     this._shouldDebug = value;
   }
 
+  /// Returns the debug value
   bool shouldDebug() {
     return this._shouldDebug;
   }
 
+  /// Returns the base API
   String getBaseApi() {
     return this._baseUrl + this._apiPath;
   }
 
-  api(RequestCallback request) {
-    return request(WPAppNetworkManager.instance);
+  /// Returns an instance of [WPAppNetworkManager] which you can use to call
+  /// your requests from.
+  api(RequestCallback request) async {
+    return await request(WPAppNetworkManager.instance);
   }
 }